added fetch favorite plants, addFavePlants, and unfavoritePlant and created a helper function in plants.js page to check if a plant has been favorited so that the users view can change based off of it. and added a helper function to hold the favorite Id so that the value can be used to remove from favorites

+++ server/api/favorite_plants.js
@@ -28,13 +28,17 @@ app.get('/', isLoggedIn,  async (req, res, next) => {
     }
 })
 
-app.delete('/:favorite_plants_id/user/user_id', isLoggedIn, async (req, res, next) => {
-    try {
+app.delete('/:favorite_plants_id/user/:user_id', isLoggedIn, async (req, res, next) => {
+try {
+    if(!req.params.favorite_plants_id ||!req.params.user_id){
+        console.log('missing fave plant or user id');
+        return res.status(400);}
+
         await deleteFavoritePlants({id: req.params.favorite_plants_id, user_id: req.params.user_id})
         res.sendStatus(204)
-    } catch (error) {
-        next(error)
-    }
+} catch (error) {
+    console.error('error in parameterized route', error)
+    next(error)
+}
 })
-
 module.exports = app;
